src/storage-service/requirements.txt
fastapi==0.115.8
motor==3.7.0
pydantic==2.10.6
pymongo==4.11.1
pytest==8.3.4
Requests==2.32.3
uvicorn==0.34.0
weaviate-client==4.11.0
